+++ hackage-security/ChangeLog.md
@@ -4,6 +4,7 @@ See also http://pvp.haskell.org/faq
 -------
 
 * Remove `Hackage.Security.TUF.FileMap.lookupM`
+* Don't expose `Hackage.Security.Util.IO` module
 * Report missing keys in `.meta` objects more appropriately as
   `ReportSchemaErrors(expected)` instead of via `Monad(fail)`
 * Add support for GHC 8.8 / base-4.13